Move UISettingsValues type to /common
