+++ libs/ui-lib/lib/common/types/ui-settings.ts
@@ -1,0 +1,5 @@
+export type UISettingsValues = {
+  addCustomManifests?: boolean;
+  customManifestsAdded?: boolean;
+  customManifestsUpdated?: boolean;
+};
