+++ libs/ui-lib/lib/ocm/components/clusterWizard/ClusterWizardContext.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import { HostsNetworkConfigurationType, UISettingsValues } from '../../services';
+import { HostsNetworkConfigurationType } from '../../services';
 import { StaticIpView } from '../clusterConfiguration/staticIp/data/dataTypes';
 import { ClusterWizardStepsType } from './wizardTransition';
+import { UISettingsValues } from '../../../common';
 
 export type ClusterWizardContextType = {
   currentStepId: ClusterWizardStepsType;
